requirements.txt
PyYAML==5.4.1
click==8.0.1
delegator.py==0.1.1
pydantic==1.8.2
whoosh==2.7.4
rich==10.6.0
